+++ src/Controller/PoorManCronController.php
@@ -89,7 +89,7 @@ class PoorManCronController
     protected function getConfigIds(string $interval): array
     {
         $models = $this->modelUtil->findModelInstancesBy('tl_entity_import_config',
-            ['tl_entity_import_config.useCron=?', 'tl_entity_import_config.cronInterval=?'], [true, $interval]);
+            ['tl_entity_import_config.useCron=?', 'tl_entity_import_config.cronInterval=?', 'tl_entity_import_config.usePoorMansCron=?'], [true, $interval, true]);
 
         if (null === $models) {
             return [];
@@ -104,17 +104,17 @@ class PoorManCronController
             return;
         }
 
-        if ($configModel->language) {
+        if ($configModel->cronLanguage) {
             $language = $GLOBALS['TL_LANGUAGE'];
 
-            $GLOBALS['TL_LANGUAGE'] = $configModel->language;
+            $GLOBALS['TL_LANGUAGE'] = $configModel->cronLanguage;
         }
 
         /** @var ImporterInterface $importer */
         $importer = $this->importerFactory->createInstance($configModel->id);
         $importer->run();
 
-        if ($configModel->language) {
+        if ($configModel->cronLanguage) {
             $GLOBALS['TL_LANGUAGE'] = $language;
         }
     }