+++ src/Resources/contao/languages/en/tl_entity_import_config.php
@@ -26,12 +26,14 @@ $lang['mergeIdentifierFields']['target'][1] = 'Select the target field into whic
 $lang['importMode'][0] = 'Import settings';
 $lang['importMode'][1] = 'Select the settings for the import here.';
 
-$lang['useCron'][0]      = 'Use cronjob';
-$lang['useCron'][1]      = 'Select this option to trigger the importer by cronjob.';
-$lang['cronInterval'][0] = 'Cron-Interval';
-$lang['cronInterval'][1] = 'Select the interval at which the import is to be performed.';
-$lang['cronDomain'][0] = 'Domain name';
-$lang['cronDomain'][1] = 'Enter the domain on which the cronjob is executed. This is for a better allocation in case errors occur during the cronjob.';
+$lang['useCron'][0]         = 'Use cronjob';
+$lang['useCron'][1]         = 'Select this option to trigger the importer by cronjob.';
+$lang['cronInterval'][0]    = 'Cron-Interval';
+$lang['cronInterval'][1]    = 'Select the interval at which the import is to be performed.';
+$lang['cronDomain'][0]      = 'Domain name';
+$lang['cronDomain'][1]      = 'Enter the domain on which the cronjob is executed. This is for a better allocation in case errors occur during the cronjob.';
+$lang['usePoorMansCron'][0] = 'Use as poor man\'s cronjob';
+$lang['usePoorMansCron'][1] = 'Choose this option to run the importer as a poor man\'s cronjob.';
 
 $lang['addSkipFieldsOnMerge'][0] = 'Add skip fields for merge';
 $lang['addSkipFieldsOnMerge'][1] = 'Select this option if you want to declare certain fields to not be overwritten on merge.';